chore(dictation): strip debug tracing, split stt tests to stt_tests.rs
Removes the temporary DEBUG(dictation-yeah) event traces from the Rust
dictation commands and the composer hook now that the decode-collapse,
prefix-stitch, and lone-filler guards are verified against live repros.
Moves the stt.rs unit tests and #[ignore]d decode experiments into a
sibling stt_tests.rs, dropping the file-size override for stt.rs.
